🎨 Added LeftsideView and Removed duplicated codes
